Updated Calculation - Correction
Made a correction to taking the derivative in the Cost Analysis section. I think the derivative of 1/(kC_{0A}(1-X) of the pdf did not account for multiplying the kC_{0A} in the numerator because of the chain rule. This should have cancelled out the kC_{0A} in the derivative which will lead to a dramatically different X value. I could easily be wrong so we should double-check.